requirements.txt
pandas>=2.1.0
numpy>=1.24.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
tqdm>=4.66.0
ccxt>=4.0.0
